==> R-Scripts/00-preamble.R <==
# preamble---------------------------------------------------------------

library(readr)
library(tidyverse)
library(rWCVP) 
library(rWCVPdata) #remotes::install_github("matildabrown/rWCVPdata")
library(data.table)
library(lme4)
library(lmerTest)
library(DHARMa)
library(mgcv)
library(parallel)
library(readxl)
library(stringr)
library(missForest)
library(see)
library(scales)
library(emmeans)
library(partR2)
library(MuMIn)
library(ggeffects)
library(patchwork)
library(rgbif)
library(sf)
library(hrbrthemes)
library(geosphere)
library(rnaturalearth)
library(rnaturalearthdata)
library(gridExtra)
library(performance)
library(conflicted)
library(ggnewscale)
library(geosphere)
library(robustlmm)
library(sjPlot)
library(showtext)
showtext_auto()
font_add_google("Roboto Condensed", "roboto")

conflict_prefer("select", "dplyr")
conflict_prefer("filter", "dplyr")
conflict_prefer("first", "dplyr")
conflict_prefer("filter", "dplyr")
